ai-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
torch==2.3.0
torchvision==0.18.0
opencv-python-headless==4.9.0.80
Pillow==10.3.0
numpy==1.26.4
python-multipart==0.0.9
